workers/agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[project]
name = "yourcast-worker"
version = "0.1.0"
description = "YourCast podcast generation worker"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.3.0",
    "redis>=5.0.0",
    "requests>=2.31.0",
    "google-generativeai>=0.8.0",  # For Gemini API key method
    "google-genai>=1.29.0",        # For service account + embeddings
    "trafilatura>=1.6.0",
    "feedparser>=6.0.10",
    "pydub>=0.25.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "webvtt-py>=0.4.6",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.6.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
